services/classification/requirements.txt
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
pillow==10.1.0
redis==5.0.1
minio==7.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
